requirements.txt
customtkinter
pillow
requests
pydantic
SQLAlchemy
pydantic-settings
redis
uvicorn
fastapi[standard]
python-jose[cryptography]
keyring
python-dotenv
aiosmtplib
websocket-client
asyncpg
psycopg
alembic